% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{create_instruments}
\alias{create_instruments}
\title{Create instruments}
\usage{
create_instruments(sf)
}
\arguments{
\item{sf}{SoundFont object as read by \code{read_sf2()}}
}
\description{
Unpack the inst, ibag, imod and igen structures into a (slightly)
more interesting nested list structure.
}
\details{
Definitely a work in progress
}
