% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{create_sample}
\alias{create_sample}
\title{Create an \code{audioSample} sample for an instrument from a SoundFont}
\usage{
create_sample(sf, inst, dur = NULL)
}
\arguments{
\item{sf}{SoundFont object as returned by \code{read_sf2()}}

\item{inst}{instrument index (integer) or name (string).
See \code{sf$pdta$shdr$name} for the names of all instruments}

\item{dur}{note duration in seconds. Default value of NULL indicates to 
       play the sample as is.  If a non-NULL duration is given, then 
\itemize{
\item{If \code{dur} is less than the natural length of the sample, the full 
      sample will be returned and no truncation will occur}
\item{If \code{dur} is longer than the natural length of the sample, the 
      sample will be looped using the looping start/end points given in the
      \code{sf$pdta$shdr} data for the instrument.}
}}
}
\value{
\code{audioSample} object compatible with the \code{audio} package
}
\description{
Create an \code{audioSample} sample for an instrument from a SoundFont
}
