% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{read_sf2}
\alias{read_sf2}
\title{Read the data in a SoundFont \code{sf2} file}
\usage{
read_sf2(filename)
}
\arguments{
\item{filename}{sf2 filename}
}
\value{
named list of SoundFont data:
\describe{
\item{\code{sfbk}}{Information about the SoundFont file from the \code{sfbk} chunk}
\item{\code{sdta}}{All audio data as a single numeric vector of floating point data 
     in the range [-1,1]}
\item{\code{pdta}}{Information about the presets, instruments and sample locations
      within the \code{sdta} data}
}
}
\description{
Read the data in a SoundFont \code{sf2} file
}
